ATR-359: fixed VS crashes which by removing global cached code analysis settings. Now every access happens via ServiceLocator
