+++ src/Acuminator/Acuminator.Analyzers/Refactorings/ChangeEventHandlerSignatureToGeneric/ChangeEventHandlerSignatureToGenericRefactoring.cs
@@ -34,7 +34,8 @@ namespace Acuminator.Analyzers.Refactorings.ChangeEventHandlerSignatureToGeneric
 			context.CancellationToken.ThrowIfCancellationRequested();
 
 			var semanticModel = await context.Document.GetSemanticModelAsync(context.CancellationToken).ConfigureAwait(false);
-			var pxContext = new PXContext(semanticModel.Compilation, GlobalCodeAnalysisSettings.Instance);
+			CodeAnalysisSettings codeAnalysisSettings = CodeAnalysisSettings.GetCodeAnalysisSettings();
+			var pxContext = new PXContext(semanticModel.Compilation, codeAnalysisSettings);
 
 			if (!pxContext.CodeAnalysisSettings.StaticAnalysisEnabled || !pxContext.IsPlatformReferenced)
 				return;
